% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bessel_layer_simulation}
\alias{bessel_layer_simulation}
\title{Bessel Layer simulation}
\usage{
bessel_layer_simulation(x, y, s, t, a)
}
\arguments{
\item{x}{start value of Brownian bridge}

\item{y}{end value of Brownian bridge}

\item{s}{start time of Brownian bridge}

\item{t}{end time of Brownian bridge}

\item{a}{vector/sequence of numbers}
}
\description{
Simulates a Bessel layer l for a given sequence a
}
\examples{
bessel_layer_simulation(x = 0, y = 0, s = 0, t = 1, a = seq(0.1, 0.5, 0.1))

}
