% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_SdeltaK_2_intervals}
\alias{calc_SdeltaK_2_intervals}
\title{Calculate interval: [S^{delta,2}_{2k+1}, S^{delta,2}_{2k}]}
\usage{
calc_SdeltaK_2_intervals(k, x, y, s, t, min, v)
}
\arguments{
\item{k}{integer value}

\item{x}{start value of Brownian bridge}

\item{y}{end value of Brownian bridge}

\item{s}{start value of Brownian bridge}

\item{t}{end value of Brownian bridge}

\item{min}{minimum of Brownian bridge}

\item{v}{upper bound of Brownian bridge}
}
\value{
vector of two values, S^{delta,2}_{2k+1} and S^{delta,2}_{2k}
}
\description{
This function calculates the interval [S^{delta,2}_{2k+1}, S^{delta, 2}_{2k}] (case where min(x,y) == min)
}
\examples{
K = ceiling(sqrt((1)+(abs(1-(-2))*abs(1-(-2))))/(2*abs(1-(-2))))
calc_SdeltaK_2_intervals(k = K, x = -2, y = 0, s = 0, t = 0, min = -2, v = 1)

}
