% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{delta_coin}
\alias{delta_coin}
\title{Delta coin flipper}
\usage{
delta_coin(x, y, s, t, min, v, k)
}
\arguments{
\item{x}{start value of Brownian bridge}

\item{y}{end value of Brownian bridge}

\item{s}{start value of Brownian bridge}

\item{t}{end value of Brownian bridge}

\item{min}{minimum of Brownian bridge}

\item{v}{upper bound of Brownian bridge}

\item{k}{integer value}
}
\value{
boolean value: if T, accept probability that Brownian bridge with minimum, min, remains in [l,v], otherwise reject
}
\description{
Flips 'Delta coin'; uses the Cauchy sequence S^{delta}_{k} to 
determine whether or not the Brownian bridge with minimum, min, 
starting at x, ending at y, between [s,t] remains in interval [l,v]
}
\examples{

}
