% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eachi}
\alias{eachi}
\title{Chi (Equation 156 and 163 in ST329)}
\usage{
eachi(j, xoy, s, t, min, v)
}
\arguments{
\item{j}{real value}

\item{xoy}{maximum of x and y where x and y are the start and end values of Brownian bridge}

\item{s}{start value of Brownian bridge}

\item{t}{end value of Brownian bridge}

\item{min}{minimum of Brownian bridge}

\item{v}{upper bound of Brownian bridge}
}
\value{
real value: chi evaluated at point j
}
\description{
This function evaluates the chi function used to simulate Bessel Layers in the infinite sums
}
\examples{
eachi(j = 1, xoy = 0, s = 0, t = 1, min = -2, v = 1)
}
