% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eadelta}
\alias{eadelta}
\title{Delta (Corollary 3 and 4 in ST329)}
\usage{
eadelta(n, x, y, s, t, min, v)
}
\arguments{
\item{n}{integer value}

\item{x}{start value of Brownian bridge}

\item{y}{end value of Brownian bridge}

\item{s}{start value of Brownian bridge}

\item{t}{end value of Brownian bridge}

\item{min}{minimum of Brownian bridge}

\item{v}{upper bound of Brownian bridge}
}
\value{
real value: S_{n}^{delta} evaluated at n
}
\description{
This function evaluates the delta function, S_{n}^{delta}
used to simulate Bessel Layers in the infinite sums
}
\examples{
# example where min(x,y) > min (delta_1 case)
eadelta(n = 1, x = 0,  y = 0.8, s = 0, t = 1, min = -2, v = 2)
eadelta1(n = 1, x = 0, y = 0.8, s = 0, t = 1, min = -2, v = 2)
# example where min(x,y) == min
eadelta(n = 1, x = -2, y = 0.8, s = 0, t = 1, min = -2, v = 2)
eadelta2(n = 1, x = -2, y = 0.8, s = 0, t = 1, min = -2, v = 2)

}
