% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eadelta1}
\alias{eadelta1}
\title{Delta_1 (Corollary 3 in ST329)}
\usage{
eadelta1(n, x, y, s, t, min, v)
}
\arguments{
\item{n}{integer value}

\item{x}{start value of Brownian bridge}

\item{y}{end value of Brownian bridge}

\item{s}{start value of Brownian bridge}

\item{t}{end value of Brownian bridge}

\item{min}{minimum of Brownian bridge}

\item{v}{upper bound of Brownian bridge}
}
\value{
real value: S_{n}^{delta,1} evaluated at n
}
\description{
This function evaluates the delta_1 function, S_{n}^{delta,1}
used to simulate Bessel Layers in the infinite sums
}
\examples{
eadelta1(n = 1, x = 0, y = 0.8, s = 0, t = 1, min = -2, v = 2)

}
