% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eadelta1_intervals}
\alias{eadelta1_intervals}
\title{Calculate interval: [S^{delta,1}_{2k+1}, S^{delta,1}_{2k}] (Corollary 3 in ST329)}
\usage{
eadelta1_intervals(k, x, y, s, t, min, v)
}
\arguments{
\item{k}{integer value}

\item{x}{start value of Brownian bridge}

\item{y}{end value of Brownian bridge}

\item{s}{start value of Brownian bridge}

\item{t}{end value of Brownian bridge}

\item{min}{minimum of Brownian bridge}

\item{v}{upper bound of Brownian bridge}
}
\value{
vector of two values, S^{delta,1}_{2k+1} and S^{delta,1}_{2k}
}
\description{
This function calculates the interval [S^{delta,1}_{2k+1}, S^{delta, 1}_{2k}] (case where min(x,y) > min)
}
\examples{
eadelta1_intervals(k = 1, x = 0, y = 0, s = 0, t = 1, min = -2, v = 1)

}
