% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eadelta2}
\alias{eadelta2}
\title{Delta_2 (Corollary 4 in ST329)}
\usage{
eadelta2(n, x, y, s, t, min, v)
}
\arguments{
\item{n}{integer value}

\item{x}{start value of Brownian bridge}

\item{y}{end value of Brownian bridge}

\item{s}{start value of Brownian bridge}

\item{t}{end value of Brownian bridge}

\item{min}{minimum of Brownian bridge}

\item{v}{upper bound of Brownian bridge}
}
\value{
real value: S_{n}^{delta,2} evaluated at n
}
\description{
This function evaluates the delta_2function, S_{n}^{delta,2}
used to simulate Bessel Layers in the infinite sums
}
\examples{
eadelta2(n = 1, x = -2, y = 0.8, s = 0, t = 1, min = -2, v = 2)

}
