% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eadelta_intervals}
\alias{eadelta_intervals}
\title{Calculate interval: [S^{delta}_{2k+1}, S^{delta}_{2k}] (Algorithm 28 in ST329)}
\usage{
eadelta_intervals(k, x, y, s, t, min, v)
}
\arguments{
\item{k}{integer value}

\item{x}{start value of Brownian bridge}

\item{y}{end value of Brownian bridge}

\item{s}{start value of Brownian bridge}

\item{t}{end value of Brownian bridge}

\item{min}{minimum of Brownian bridge}

\item{v}{upper bound of Brownian bridge}
}
\value{
vector of two values, S^{delta}_{2k+1} and S^{delta}_{2k}
}
\description{
This function calculates the interval [S^{delta}_{2k+1}, S^{delta}_{2k}] 
(case where min(x,y) > min or where min(x,y) == min)
}
\examples{
# case where min(x,y ) > min
eadelta1_intervals(k = 1, x = 0, y = 0, s = 0, t = 1, min = -2, v = 1)

# case where min(x,y) == min
K = ceiling(sqrt((1)+(abs(1-(-2))*abs(1-(-2))))/(2*abs(1-(-2))))
eadelta_intervals(k = K, x = -2, y = 0, s = 0, t = 0, min = -2, v = 1)

}
