% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eagamma}
\alias{eagamma}
\title{Gamma (Corollary 2 and Algorithm 26 in ST329)}
\usage{
eagamma(n, x, y, s, t, l, v)
}
\arguments{
\item{n}{integer value}

\item{x}{start value of Brownian bridge}

\item{y}{end value of Brownian bridge}

\item{s}{start value of Brownian bridge}

\item{t}{end value of Brownian bridge}

\item{l}{lower bound of Brownian bridge}

\item{v}{upper bound of Brownian bridge}
}
\value{
real value: S_{n}^{gamma} evaluated at n
}
\description{
This function evaluates the gamma function, S_{n}^{gamma},
used to simulate Bessel Layers in the infinite sums
}
\examples{
eagamma(n = 1, x = 0, y = 0, s = 0, t = 1, l = -0.4, v = 0.8)
}
