% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eagamma_intervals}
\alias{eagamma_intervals}
\title{Calculate interval: [S^{gamma}_{2k+1}, S^{gamma}_{2k}] (Corollary 2 and Algorithm 26 in ST329)}
\usage{
eagamma_intervals(k, x, y, s, t, l, v)
}
\arguments{
\item{k}{integer value}

\item{x}{start value of Brownian bridge}

\item{y}{end value of Brownian bridge}

\item{s}{start value of Brownian bridge}

\item{t}{end value of Brownian bridge}

\item{l}{lower bound of Brownian bridge}

\item{v}{upper bound of Brownian bridge}
}
\value{
vector of two values, S^{gamma}_{2k+1} and S^{gamma}_{2k}
}
\description{
This function calculates the interval [S^{gamma}_{2k+1}, S^{gamma}_{2k}] for given k
}
\examples{
eagamma_intervals(k = 1, x = 0, y = 0, s = 0, t = 1, l = -2, v = 1)
}
