% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{easigma}
\alias{easigma}
\title{Sigma (Equation 140 in ST329)}
\usage{
easigma(j, x, y, s, t, l, v)
}
\arguments{
\item{j}{real value}

\item{x}{start value of Brownian bridge}

\item{y}{end value of Brownian bridge}

\item{s}{start value of Brownian bridge}

\item{t}{end value of Brownian bridge}

\item{l}{lower bound of Brownian bridge}

\item{v}{upper bound of Brownian bridge}
}
\value{
real value: sigma evaluated at point j
}
\description{
This function evaluates the sigma function used to simulate Bessel Layers in the infinite sums
}
\examples{
easigma(j = 1, x = 0, y = 0, s = 0, t = 1, l = -2, v = 1)

}
