% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_max}
\alias{find_max}
\title{Find maximum of vector}
\usage{
find_max(vect)
}
\arguments{
\item{vect}{numerical vector}
}
\value{
maximum of vector given
}
\description{
This function finds the maximum value of a numerical vector
}
\examples{
# returns 0.8
find_max(c(-0.4, 0.23, -0.4321, 0.6, 0.3, 0.8, 0.54)) 

}
