% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_min}
\alias{find_min}
\title{Find minimum of vector}
\usage{
find_min(vect)
}
\arguments{
\item{vect}{numerical vector}
}
\value{
minimum of vector given
}
\description{
This function finds the minimum value of a numerical vector
}
\examples{
# returns -0.4321
find_min(c(-0.4, 0.23, -0.4321, 0.6, 0.3, 0.8, 0.54)) 

}
