% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gamma_coin}
\alias{gamma_coin}
\title{Gamma coin flipper}
\usage{
gamma_coin(x, y, s, t, l, v, k)
}
\arguments{
\item{x}{start value of Brownian bridge}

\item{y}{end value of Brownian bridge}

\item{s}{start value of Brownian bridge}

\item{t}{end value of Brownian bridge}

\item{l}{lower bound of Brownian bridge}

\item{v}{upper bound of Brownian bridge}

\item{k}{integer value}
}
\value{
boolean value: if T, accept probability that Brownian bridge remains in [l,v], otherwise reject
}
\description{
Flips 'Gamma coin'; uses the Cauchy sequence S^{gamma}_{k} to 
determine whether or not the Brownian bridge starting at x, ending at y, between [s,t]
remains in interval [l,v]
}
\examples{
gamma_coin(x = 0, y = 0, s = 0, t = 1, l = -0.5, v = 0.5, k = 1)

}
