\name{layeredBB-package}
\alias{layeredBB-package}
\alias{layeredBB}
\docType{package}
\title{
  Simulates layered Brownian bridges using C++ and Rcpp
}
\description{
  R access to C++ implementation of layered Brownian Bridge simulation and Bessel layer simulation.
}
\details{
  This package includes functions to simulated layered Brownian bridges using Bessel layers. See Pollock et al. (2016) and Pollock (2013) for more details on simulating Bessel layers using Cauchy sums. 
}
\author{
  Ryan Chan, rchan@turing.ac.uk.
}
\references{
  M. Pollock. Some Monte Carlo Methods for Jump Diffusions. PhD thesis, University of Warwick, September 2013. 
  
  M.Pollock, A.M. Johansen, G.O. Roberts. 2016. On the Exact and epsilon-Strong Simulation of (Jump) Diffusions. Bernoulli. 22(2), pp. 794-856. 
  
  Beskos, A., Papaspiliopoulos, O., Roberts, G.O., and Fearnhead, P. 2006. Exact and computationally effcient likelihood-based estimation for discretely observed diffusion processes (with discussion). Journal of Royal Statistical Society, B. 68, pp. 333-382.
}
\keyword{ package }
