% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{M_function}
\alias{M_function}
\title{M_function}
\usage{
M_function(a, x, y, s, t)
}
\arguments{
\item{a}{real value}

\item{x}{start value of Brownian bridge}

\item{y}{end value of Brownian bridge}

\item{s}{start value of Brownian bridge}

\item{t}{end value of Brownian bridge}
}
\value{
real value: M_function evaluated at point a
}
\description{
This function is used in the simulation of a minimum point of a Brownian bridge
}
\examples{
M_function(a = 0, x = 0, y = 0, s = 0, t = 1)

}
