% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{max_sampler}
\alias{max_sampler}
\title{Brownian Bridge maximum point sampler}
\usage{
max_sampler(x, y, s, t, low_bound, up_bound)
}
\arguments{
\item{x}{start value of Brownian bridge}

\item{y}{end value of Brownian bridge}

\item{s}{start value of Brownian bridge}

\item{t}{end value of Brownian bridge}

\item{low_bound}{Lower bound of maximum point (max(x,y) <= low_bound < up_bound)}

\item{up_bound}{Upper bound of maximum point (max(x,y) <= low_bound < up_bound)}
}
\value{
vector: the simulated maximum, 'max', and time where maximum occurs, 'tau'
}
\description{
This function simulates a maximum point of a Brownian bridge
}
\examples{
# simulate a maximum between 0 and 1 of a Brownian bridge starting at 0 and ending at 0 in time [0,1]
max_sampler(x=0, y=0, s=0, t=1, low_bound = 0, up_bound = 1)

}
