% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{min_Bessel_bridge_path_sampler}
\alias{min_Bessel_bridge_path_sampler}
\title{Bessel Bridge path sampler given minimum}
\usage{
min_Bessel_bridge_path_sampler(x, y, s, t, m, tau, times, checks = TRUE)
}
\arguments{
\item{x}{start value of Bessel bridge}

\item{y}{end value of Bessel bridge}

\item{s}{start value of Bessel bridge}

\item{t}{end value of Bessel bridge}

\item{m}{minumum point}

\item{tau}{time of minimum point}

\item{times}{vector of real numbers to simulate Bessel bridge}

\item{checks}{logical value to determine if arguments that are passed into 
the function are checked. Things that are checked include that 
s < t, that requested simulation times are in [s,t], that m <= min(x,y)
and that if tau == s or tau == t, then m == x or m == y, respectively}
}
\value{
A list with the following components
\describe{
  \item{full_path}{Matrix of the simulated Bessel bridge path at all 
                   included time points, i.e. s, t and times. The times
                   are sorted and duplicates are removed. The first row
                   are the points of the Brownian bridge (named 'X') 
                   second row are corresponding times (named 'times')}
  \item{simulated_path}{Matrix of the simulated Bessel bridge path only at 
                        the specified times passed into the function, i.e. 
                        the times vector. The times are not sorted and
                        duplicates are not removed. The first row
                        are the points of the Bessel bridge (named 'X') 
                        second row are corresponding times (named 'times')}
  \item{remove_m_path}{Matrix of the simulated Bessel bridge path only at 
                       all included times points excluding tau. These times
                       are sorted and duplicates are removed. The first row
                       are the points of the Bessel bridge (named 'X') 
                       second row are corresponding times (named 'times'). 
                       Note that the minimum point is included if it is 
                       passed into the times vector}
}
}
\description{
Simulation of a path of a Bessel bridge at given times, given minimum occurs at time tau
}
\examples{
# simulating a path at times=c(0.2, 0.4, 0.8) for a Bessel bridge starting 
# at 0 and ending at 0 in time [0,1] given minimum is at -0.4 at time 0.6
min_Bessel_bridge_path_sampler(x = 0,
                               y = 0,
                               s = 0,
                               t = 1,
                               m = -0.4,
                               tau = 0.6,
                               times = c(0.2, 0.4, 0.8))

# note that remove_m_path will still include the minimum if passed into times
# also note that simulated_path does not remove duplicates passed into times
min_Bessel_bridge_path_sampler(x = 0,
                               y = 0,
                               s = 0,
                               t = 1,
                               m = -0.4,
                               tau = 0.6,
                               times = c(0.2, 0.4, 0.6, 0.8, 0.6))

# another example
start <- runif(1, -1, 1)
end <- runif(1, -1, 1)
min <- min_sampler(x = start,
                   y = end,
                   s = 0,
                   t = 1,
                   low_bound = min(start, end)-0.4,
                   up_bound = min(start, end)-0.2)
path <- min_Bessel_bridge_path_sampler(x = start,
                                       y = end,
                                       s = 0,
                                       t = 1,
                                       m = min['min'],
                                       tau = min['tau'],
                                       times = seq(0, 1, 0.01))$full_path
plot(x = path['time',], y = path['X',], pch = 20, xlab = 'Time', ylab = 'X')
lines(x = path['time',], y = path['X',])
points(x = min['tau'], y = min['min'], col = 'red', pch = 20)
}
