% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{min_sampler}
\alias{min_sampler}
\title{Brownian Bridge minimum point sampler}
\usage{
min_sampler(x, y, s, t, low_bound, up_bound)
}
\arguments{
\item{x}{start value of Brownian bridge}

\item{y}{end value of Brownian bridge}

\item{s}{start value of Brownian bridge}

\item{t}{end value of Brownian bridge}

\item{low_bound}{Lower bound of minimum point (low_bound < up_bound <= min(x,y))}

\item{up_bound}{Upper bound of minimum point (low_bound < up_bound <= min(x,y))}
}
\value{
vector: the simulated minimum, 'min', and time where minimum occurs, 'tau'
}
\description{
This function simulates a minimum point of a Brownian bridge
}
\examples{
# simulate a minimum between 0 and 1 of a Brownian bridge starting at 0 and ending at 0 in time [0,1]
min_sampler(x=0, y=0, s=0, t=1, low_bound = -1, up_bound = 0)

}
