% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multi_bessel_layer_simulation}
\alias{multi_bessel_layer_simulation}
\title{Multiple Bessel Layer simulation}
\usage{
multi_bessel_layer_simulation(dim, x, y, s, t, a)
}
\arguments{
\item{dim}{dimension of Brownian bridge}

\item{x}{vector of start values of Brownian bridge (length of vector must be equal to dim)}

\item{y}{vector of end values of Brownian bridge (length of vector must be equal to dim)}

\item{s}{start time of Brownian bridge}

\item{t}{end time of Brownian bridge}

\item{a}{vector/sequence of numbers}
}
\value{
A list of length dim where list[i] is the Bessel layer for component i, which is represented in
a list with the following items:
\itemize{
  \item a vector of real values
  \item l bessel layer
}
where the Bessel layer [min(x,y)-a[l], min(x,y)+a[l]]
}
\description{
Simulates a Bessel layer l for a given sequence a for each component of the Brownian bridge
}
\examples{
# simulate layer information for two-dimensional Brownian bridge starting and ending at (0,0) in time [0,1]
multi_bessel_layer_simulation(dim = 2, x = c(0, 0), y = c(0, 0), s = 0, t = 1, a = seq(0.1, 0.5, 0.1))

}
