% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multi_brownian_bridge}
\alias{multi_brownian_bridge}
\title{Multi-dimensional Brownian Bridge path sampler}
\usage{
multi_brownian_bridge(dim, x, y, s, t, times)
}
\arguments{
\item{dim}{dimension of Brownian bridge}

\item{x}{start value of Brownian bridge}

\item{y}{end value of Brownian bridge}

\item{s}{start value of Brownian bridge}

\item{t}{end value of Brownian bridge}

\item{times}{vector of real numbers to simulate Brownian bridge}
}
\value{
matrix of the simulated layered Brownian bridge path, 
        first dim rows are points for X in each component, 
        last row are corresponding times
}
\description{
Simulation of a multi-dimensional Brownian Bridge, at given times
}
\examples{
# simulate two-dimensional Brownian bridge starting and ending 
# at (0,0) in time [0,1]
multi_brownian_bridge(dim = 2,
                      x = c(0,0),
                      y = c(0,0),
                      s = 0,
                      t = 1,
                      times = seq(0.2, 0.8, 0.2))

}
