% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multi_brownian_motion}
\alias{multi_brownian_motion}
\title{Multi-dimensional Brownian Motion path sampler}
\usage{
multi_brownian_motion(dim, x, times)
}
\arguments{
\item{dim}{dimension of Brownian motion}

\item{x}{start value of Brownian motion}

\item{times}{vector of real numbers to simulate Brownian motion
(must be a vector of at least length two, where times[1]
 is the start time of the Brownian motion, and times[2:length(times)]
 are the times you wish to further simulate)}
}
\value{
Matrix of the simulated Brownian motion path at all
        included time points. The times are sorted. 
        The first dim rows are the points of the Brownian motion
        dim+1 row are corresponding times
}
\description{
Simulation of a multi-dimensional Brownian Motion, at given times
}
\examples{
# simulate two-dimensional Brownian bridge starting and ending 
# at (0,0) in time [0,1]
multi_brownian_motion(dim = 2,
                      x = c(0,0),
                      times = c(0.1, 0.2, 0.4, 0.6, 0.8))
                      
# note that the times are sorted and duplicates are removed
multi_brownian_motion(dim = 2,
                      x = c(0.5,1),
                      times = c(0.1, 0.2, 0.4, 0.6, 0.6, 0.8, 0.1))
}
