% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{product_vector_elements}
\alias{product_vector_elements}
\title{Find product of a vector}
\usage{
product_vector_elements(vect)
}
\arguments{
\item{vect}{numerical vector}
}
\value{
product of element in vector given
}
\description{
This function product of the elements of a numerical vector
}
\examples{
# returns 120
find_max(c(1,2,3,4,5)) 

}
