% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{psi}
\alias{psi}
\title{Psi}
\usage{
psi(j, x, y, s, t, min, v)
}
\arguments{
\item{j}{real value}

\item{x}{start value of Brownian bridge}

\item{y}{end value of Brownian bridge}

\item{s}{start value of Brownian bridge}

\item{t}{end value of Brownian bridge}

\item{min}{minimum of Brownian bridge}

\item{v}{upper bound of Brownian bridge}
}
\value{
real value: psi evaluated at point j
}
\description{
This function evaluates the psi function used to simulate Bessel Layers in the infinite sums
}
\examples{
psi(j = 1, x = 0, y = 0, s = 0, t = 1, min = -2, v = 1)

}
