% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sigma_bar}
\alias{sigma_bar}
\title{Sigma_Bar}
\usage{
sigma_bar(j, x, y, s, t, l, v)
}
\arguments{
\item{j}{real value}

\item{x}{start value of Brownian bridge}

\item{y}{end value of Brownian bridge}

\item{s}{start value of Brownian bridge}

\item{t}{end value of Brownian bridge}

\item{l}{lower bound of Brownian bridge}

\item{v}{upper bound of Brownian bridge}
}
\value{
real value: sigma_bar evaluated at point j
}
\description{
This function evaluates the sigma_bar function used to simulate Bessel Layers in the infinite sums
}
\examples{
sigma_bar(j = 1, x = 0, y = 0, s = 0, t = 1, l = -2, v = 1)

}
