% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bessel_layer_simulation}
\alias{bessel_layer_simulation}
\title{Bessel Layer simulation}
\usage{
bessel_layer_simulation(x, y, s, t, mult = 1)
}
\arguments{
\item{x}{start value of Brownian bridge}

\item{y}{end value of Brownian bridge}

\item{s}{start time of Brownian bridge}

\item{t}{end time of Brownian bridge}

\item{a}{vector/sequence of numbers}
}
\value{
A list with the following items:
\describe{
  \item{L}{Hard lower bound}
  \item{l}{Soft lower bound}
  \item{u}{Soft upper bound}
  \item{U}{Hard upper bound}
}
where the Bessel layer is [L, U] and either the minimum occurs in [L, l] or
the maximum occurs in [u, U]
}
\description{
Simulates a Bessel layer l for a given sequence a
}
\examples{
bessel_layer_simulation(x = 0,
                        y = 0,
                        s = 0,
                        t = 1,
                        mult = 0.5)

}
