% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Brownian_bridge_path_sampler}
\alias{Brownian_bridge_path_sampler}
\title{Brownian Bridge path sampler}
\usage{
Brownian_bridge_path_sampler(x, y, s, t, times)
}
\arguments{
\item{x}{start value of Brownian bridge}

\item{y}{end value of Brownian bridge}

\item{s}{start value of Brownian bridge}

\item{t}{end value of Brownian bridge}

\item{times}{vector of real numbers to simulate Brownian bridge}
}
\value{
matrix of the simulated Brownian bridge path, first row is points X, second row are corresponding times
}
\description{
This function simulates a path of a Brownian bridge at given times
}
