% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{min_Bessel_bridge_sampler}
\alias{min_Bessel_bridge_sampler}
\title{Bessel Bridge point sampler given minimum (Algorithm 15 in ST329)}
\usage{
min_Bessel_bridge_sampler(x, y, s, t, min, tau, q)
}
\arguments{
\item{x}{start value of Bessel bridge}

\item{y}{end value of Bessel bridge}

\item{s}{start value of Bessel bridge}

\item{t}{end value of Bessel bridge}

\item{min}{minumum point}

\item{tau}{time of minimum point}

\item{q}{time of simulation}
}
\value{
simulated point of the Bessel bridge at time q
}
\description{
Simulation of a point of a Bessel bridge at time q, given minimum occurs at time tau
}
\examples{
# simulating a point at q=0.2 for a Bessel bridge starting at 0 and ending 
# at 0 in time [0,1] given minimum is at -0.4 at time 0.6
min_Bessel_bridge_sampler(x = 0,
                          y = 0,
                          s = 0,
                          t = 1,
                          min = -0.4,
                          tau = 0.6,
                          q = 0.2)

}
